+++ src/detection/packages/packages.h
@@ -6,6 +6,7 @@
 typedef struct FFPackagesResult {
     uint32_t amSystem;
     uint32_t amUser;
+    uint32_t appimage;
     uint32_t apk;
     uint32_t brew;
     uint32_t brewCask;